Refactor InvitesPage to extract InviteItem component
